+++ src/vendor_api/pgs.py
@@ -44,7 +44,7 @@ class PGSAPI(VendorAPI):
                 card_listing = Card_Listing(
                     card_spec=listing_card_spec,
                     store="PGS",
-                    price=int(card_listing['price']),
+                    price=float(card_listing['price']),
                     currency="AUD",
                     price_unit="AUD",
                     quantity=card_listing['quantity'],
